requirements.txt
Django>=5.2,<6.1
Pillow>=10.4
gunicorn>=22.0
whitenoise>=6.7
dj-database-url>=2.2
psycopg[binary]>=3.2
python-dotenv>=1.0
